downstream: arm64/mpam: fix impossible condition in resctrl_arch_rmid_read()
The smatch tool complains about impossible condition warning shown below:

drivers/platform/mpam/mpam_resctrl.c:414 resctrl_arch_rmid_read()
warn: impossible condition '(cfg.mon == ((~0) + 1)) => (0-u16max == 65536)'

The USE_RMID_IDX macro is a 32-bit data type, as the assignment object,
cfg.mon should also be a 32-bit unsigned variable.

Fixes: 21ac9fc8f275 ("arm_mpam: Track bandwidth counter state for overflow and power management")

+++ drivers/platform/mpam/mpam_internal.h
@@ -190,7 +190,7 @@ enum mon_filter_options {
 };
 
 struct mon_cfg {
-	u16                     mon;
+	u32                     mon;
 	u8                      pmg;
 	bool                    match_pmg;
 	u32                     partid;
